@@ -19,7 +19,7 @@
     }
   },
   "require": {
-    "php": "^7.0|^7.1",
+    "php": "^8.1",
     "paragonie/csp-builder": "^2.0"
   },
   "autoload-dev": {
@@ -28,10 +28,17 @@
     }
   },
   "require-dev": {
-    "friendsofphp/php-cs-fixer": "^2.3",
+    "friendsofphp/php-cs-fixer": "^3",
     "psr/http-message": "^1.0",
-    "phpunit/phpunit": "^6.1",
-    "container-interop/container-interop": "^1.2"
+    "phpunit/phpunit": "^9",
+    "container-interop/container-interop": "^1.2",
+    "phpstan/phpstan": "^1.9"
+  },
+  "scripts": {
+    "test": "phpunit -c tests/phpunit/phpunit.xml --colors=always",
+    "fix-styles": "php-cs-fixer fix --allow-risky=yes",
+    "check-styles": "php-cs-fixer fix --allow-risky=yes --dry-run",
+    "run-static-analysis": "phpstan analyse src/"
   },
   "config": {
     "bin-dir": "bin",
